debug-scripts/V4-FoundryProjectV2/requirements.txt
azure-ai-evaluation==1.11.0
azure-ai-inference==1.0.0b9
azure-ai-projects==2.0.0b3
azure-cosmos==4.9.0
azure-identity==1.24.0
azure-monitor-events-extension==0.1.0
azure-monitor-opentelemetry==1.8.5
azure-search-documents==11.5.3
fastapi==0.116.1
openai==2.16.0
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-exporter-otlp-proto-http==1.39.0
opentelemetry-instrumentation-fastapi==0.60b0
opentelemetry-instrumentation-openai==0.46.2
opentelemetry-sdk==1.39.0
pytest==8.4.1
pytest-asyncio==0.24.0
pytest-cov==5.0.0
python-dotenv==1.1.1
python-multipart==0.0.20
semantic-kernel==1.35.3
uvicorn==0.35.0
pylint-pydantic==0.3.5
pexpect==4.9.0
mcp==1.26.0
agent-framework-azure-ai==1.0.0b260130
agent-framework-core==1.0.0b260130
opentelemetry-semantic-conventions-ai==0.4.13
pyodbc
ipykernel
pydantic
